- Add beginnings of a Docker image/compose file - Make a setup GUI - Do some stuff.. to make the setup work? - Add a logger from Companion to Server - Fix authfulPage.js (bugs include not working, querySelector errors) - Fix HookLoader race condition(s) - Fix all sounds going on top of eachother after page switches back - Fix position setting - Workaround for browser Companion support for playing sounds (just ask permission and yell at the browser until it gives)

+++ src/index.js
@@ -5,41 +5,70 @@ const debug = require('./utils/debug');
 
 let DOES_RUN_SERVER = true;
 
+const DOES_SETTINGS_EXIST_YET = fs.existsSync(path.resolve('./src/configs/config.fd.js'));
+
+if (!DOES_SETTINGS_EXIST_YET) {
+  require(path.resolve('./src/private/setup.js'))().then(() => {
+    console.log(picocolors.bgGreen('Setup complete!'));
+    const spawn = require('child_process').spawn;
+
+    process.on('exit', () => {
+      const child = spawn('npm', ['start'], {
+        detached: true,
+        stdio: 'ignore',
+        cwd: process.cwd(),
+        shell: true,
+        env: process.env,
+      });
+
+      child.unref();
+    });
+  });
+}
+
 if (process.argv[2] === 'server') {
   console.log(picocolors.blue('Server only mode.'));
 } else {
-  require('./companionInit');
+  if (DOES_SETTINGS_EXIST_YET) require('./companionInit')('./src/fdconnect.html', true, 850, 550, false);
   if (process.argv[2] == 'companion') {
     console.log(picocolors.blue('Companion only mode.'));
     DOES_RUN_SERVER = false;
   }
 }
 
-if (DOES_RUN_SERVER) require('./server');
+if (DOES_RUN_SERVER && DOES_SETTINGS_EXIST_YET) require('./server');
 
-['SIGHUP', 'SIGINT', 'SIGQUIT', 'SIGILL', 'SIGTRAP', 'SIGABRT',
-  'SIGBUS', 'SIGFPE', 'SIGUSR1', 'SIGSEGV', 'SIGUSR2', 'SIGTERM', 'exit',
-].forEach((sig) => {
-  process.on(sig, () => {
-    if (sig == 'SIGINT') {
-      console.log('So you have chosen an unclean death. Goodbye.');
-    }
-    terminator(sig);
-    debug.log('Signal received: ' + sig);
-  });
-});
 
-const terminator = (sig) => {
-  if (typeof sig === 'string') {
-    // call your async task here and then call process.exit() after async task is done
-    fs.rmSync(path.resolve('./src/public/hooks'), {recursive: true});
-    console.log('Hooks unloaded.');
-    fs.rmSync(path.resolve('./tmp'), {recursive: true});
-    console.log('Plugins unloaded. Bye!');
-
-    setTimeout(() => {
-      process.exit(1);
+const setupTerm = () => {
+  ['SIGHUP', 'SIGINT', 'SIGQUIT', 'SIGILL', 'SIGTRAP', 'SIGABRT',
+    'SIGBUS', 'SIGFPE', 'SIGUSR1', 'SIGSEGV', 'SIGUSR2', 'SIGTERM', 'exit',
+  ].forEach((sig) => {
+    process.on(sig, () => {
+      if (sig == 'SIGINT') {
+        console.log('So you have chosen an unclean death. Goodbye.');
+      }
+      terminator(sig);
+      debug.log('Signal received: ' + sig);
     });
-  }
+  });
+
+  const terminator = (sig) => {
+    if (typeof sig === 'string') {
+    // call your async task here and then call process.exit() after async task is done
+      if (fs.existsSync(path.resolve('./src/public/hooks'))) {
+        fs.rmSync(path.resolve('./src/public/hooks'), {recursive: true});
+        console.log('Hooks unloaded.');
+      }
+      if (fs.existsSync(path.resolve('./tmp'))) {
+        fs.rmSync(path.resolve('./tmp'), {recursive: true});
+        console.log('Plugins unloaded. Bye!');
+      }
+
+      setTimeout(() => {
+        process.exit(1);
+      });
+    }
+  };
 };
 
+if (DOES_SETTINGS_EXIST_YET) setupTerm();
